RAG/indexing/requirements.txt
qdrant-client==1.10.1
requests>=2.31.0
pydantic>=2.8.0
PyYAML>=6.0.1
click>=8.1.7
tqdm>=4.66.4
# Optional local embeddings (fallback)
sentence-transformers>=3.0.1
# Optional document parsing fallbacks
pymupdf>=1.24.5
python-docx>=1.1.2
